Add middlewares and auth

+++ controllers/cards.js
@@ -17,7 +17,16 @@ module.exports.createCard = (req, res) => {
 
 module.exports.deleteCard = (req, res) => {
   const { cardId } = req.params;
-  Card.findByIdAndDelete(cardId)
-    .then(() => res.status(200).send({ data: cardId }))
-    .catch((err) => res.status(500).send({ message: err.message }));
-}
+  const owner = req.user._id;
+
+  Card.findById(cardId)
+    .then((card) => {
+      if (card.owner.toString() === owner) {
+        Card.findByIdAndDelete(cardId)
+          .then(() => res.status(200).send({ data: cardId }))
+          .catch((err) => res.status(500).send({ message: err.message }));
+      } else {
+        return res.status(403).send({ message: 'У вас нет прав для удаления данной карточки' });
+      }
+    }).catch((err) => res.status(404).send({ message: err }));
+};
